+++ src/hopfield_model.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-HopfieldModel::HopfieldModel(const bool _sync, const double _potential, const string& weights_file, const string& bias_file,
+HopfieldModel::HopfieldModel(const bool _sync, const float _potential, const string& weights_file, const string& bias_file,
 		const string& output_file, const int seed): 
 		NeuralNetworkModel(weights_file, bias_file, output_file, seed),sync(_sync),potential(_potential){
   
@@ -15,14 +15,14 @@ void HopfieldModel::updateSync(){
   for(uint32_t i=0; i<neurons.size(); ++i){
 
     //calc input sum
-    double input_sum = 0;
+    float input_sum = 0;
     for(const auto& w : weights[i]){
       input_sum += neurons[w.neuron_id].getOutput() * w.weight;
     }
     input_sum += neurons[i].getBias();
 
     //calc output
-    double output = func(input_sum);
+    float output = func(input_sum);
 
     //set output
     neurons[i].setOutput(output);
@@ -39,14 +39,14 @@ void HopfieldModel::updateAsync(){
     int id = rand_int(mt);
 
     //calc input sum
-    double input_sum = 0;
+    float input_sum = 0;
     for(const auto& w : weights[id]){
       input_sum += neurons[w.neuron_id].getOutput() * w.weight;
     }
     input_sum += neurons[id].getBias();
 
     //calc output
-    double output = func(input_sum);
+    float output = func(input_sum);
 
     //set output
     neurons[id].setOutput(output);
@@ -65,6 +65,6 @@ void HopfieldModel::update(){
   }
 }
 
-double HopfieldModel::func(double input_sum){
+float HopfieldModel::func(float input_sum){
 	return (std::tanh(input_sum / potential) + 1) / 2;
 }
